Design for new player api
