+++ gonk-player/Cargo.toml
@@ -16,6 +16,7 @@ symphonia = { git = "https://github.com/pdeljanov/Symphonia", default-features =
     "ogg",
     "vorbis",
 ] }
+ringbuf = "*"
 
 [target.'cfg(target_os = "windows")'.dependencies]
 winapi = { version = "0.3.9", features = [
