+++ gonk-player/src/lib.rs
@@ -2,27 +2,14 @@
     clippy::not_unsafe_ptr_arg_deref,
     clippy::missing_safety_doc,
     non_upper_case_globals,
-    non_snake_case
+    non_snake_case,
+    clippy::type_complexity
 )]
 use crossbeam_channel::{bounded, Sender};
 use gonk_core::{Index, Song};
-use std::fs::File;
-use std::io::ErrorKind;
-use std::{collections::VecDeque, thread, time::Duration};
-use symphonia::core::errors::Error;
-use symphonia::core::formats::{FormatReader, Track};
-use symphonia::{
-    core::{
-        audio::SampleBuffer,
-        codecs::{Decoder, DecoderOptions},
-        formats::{FormatOptions, SeekMode, SeekTo},
-        io::MediaSourceStream,
-        meta::MetadataOptions,
-        probe::Hint,
-        units::Time,
-    },
-    default::get_probe,
-};
+use std::{thread, time::Duration};
+
+pub mod decoder;
 
 #[cfg(windows)]
 mod wasapi;
@@ -207,143 +194,5 @@ impl Player {
         self.s
             .send(Event::OutputDevice(device.to_string()))
             .unwrap();
-    }
-}
-
-pub struct Symphonia {
-    format_reader: Box<dyn FormatReader>,
-    decoder: Box<dyn Decoder>,
-    track: Track,
-    elapsed: u64,
-    duration: u64,
-    error_count: u8,
-    buf: VecDeque<f32>,
-}
-
-impl Symphonia {
-    pub fn new(path: &str) -> Result<Self, Box<dyn std::error::Error>> {
-        let file = File::open(path)?;
-        let mss = MediaSourceStream::new(Box::new(file), Default::default());
-        let probed = get_probe().format(
-            &Hint::default(),
-            mss,
-            &FormatOptions {
-                prebuild_seek_index: true,
-                seek_index_fill_rate: 1,
-                enable_gapless: false,
-            },
-            &MetadataOptions::default(),
-        )?;
-
-        let track = probed.format.default_track().ok_or("")?.to_owned();
-        let n_frames = track.codec_params.n_frames.ok_or("")?;
-        let duration = track.codec_params.start_ts + n_frames;
-        let decoder = symphonia::default::get_codecs()
-            .make(&track.codec_params, &DecoderOptions::default())?;
-
-        Ok(Self {
-            format_reader: probed.format,
-            decoder,
-            track,
-            duration,
-            elapsed: 0,
-            error_count: 0,
-            buf: VecDeque::new(),
-        })
-    }
-    pub fn elapsed(&self) -> Duration {
-        let tb = self.track.codec_params.time_base.unwrap();
-        let time = tb.calc_time(self.elapsed);
-        Duration::from_secs(time.seconds) + Duration::from_secs_f64(time.frac)
-    }
-    pub fn duration(&self) -> Duration {
-        let tb = self.track.codec_params.time_base.unwrap();
-        let time = tb.calc_time(self.duration);
-        Duration::from_secs(time.seconds) + Duration::from_secs_f64(time.frac)
-    }
-    pub fn sample_rate(&self) -> u32 {
-        self.track.codec_params.sample_rate.unwrap()
-    }
-    //TODO: I would like seeking out of bounds to play the next song.
-    //I can't trust symphonia to provide accurate errors so it's not worth the hassle.
-    //I could use pos + elapsed > duration but the duration isn't accurate.
-    pub fn seek(&mut self, pos: f32) {
-        let pos = Duration::from_secs_f32(pos);
-
-        //Ignore errors.
-        let _ = self.format_reader.seek(
-            SeekMode::Coarse,
-            SeekTo::Time {
-                time: Time::new(pos.as_secs(), pos.subsec_nanos() as f64 / 1_000_000_000.0),
-                track_id: None,
-            },
-        );
-    }
-    #[allow(clippy::should_implement_trait)]
-    pub fn next(&mut self) -> Option<f32> {
-        if self.buf.is_empty() {
-            match self.next_packet() {
-                Some(packet) => self.buf = VecDeque::from(packet.samples().to_vec()),
-                None => {
-                    return None;
-                }
-            }
-        }
-
-        self.buf.pop_front()
-    }
-    pub fn next_packet(&mut self) -> Option<SampleBuffer<f32>> {
-        if self.error_count > 2 || unsafe { &STATE } == &State::Finished {
-            return None;
-        }
-
-        let next_packet = match self.format_reader.next_packet() {
-            Ok(next_packet) => {
-                self.error_count = 0;
-                next_packet
-            }
-            Err(err) => match err {
-                Error::IoError(e) if e.kind() == ErrorKind::UnexpectedEof => {
-                    //Just in case my 250ms addition is not enough.
-                    if self.elapsed() + Duration::from_secs(1) > self.duration() {
-                        unsafe { STATE = State::Finished };
-                        return None;
-                    } else {
-                        self.error_count += 1;
-                        return self.next_packet();
-                    }
-                }
-                _ => {
-                    gonk_core::log!("{}", err);
-                    self.error_count += 1;
-                    return self.next_packet();
-                }
-            },
-        };
-
-        self.elapsed = next_packet.ts();
-        unsafe { ELAPSED = self.elapsed() };
-
-        //HACK: Sometimes the end of file error does not indicate the end of the file?
-        //The duration is a little bit longer than the maximum elapsed??
-        //The final packet will make the elapsed time move backwards???
-        if self.elapsed() + Duration::from_millis(250) > self.duration() {
-            unsafe { STATE = State::Finished };
-            return None;
-        }
-
-        match self.decoder.decode(&next_packet) {
-            Ok(decoded) => {
-                let mut buffer =
-                    SampleBuffer::<f32>::new(decoded.capacity() as u64, *decoded.spec());
-                buffer.copy_interleaved_ref(decoded);
-                Some(buffer)
-            }
-            Err(err) => {
-                gonk_core::log!("{}", err);
-                self.error_count += 1;
-                self.next_packet()
-            }
-        }
     }
 }
