+++ gonk-player/src/main.rs
@@ -1,0 +1,5 @@
+use gonk_player::{decoder::*, *};
+
+fn main() {
+    let decoder = Decoder::new("test");
+}
